finish component coursedetail rectif course controller in server add profil component TODO

+++ todo/src/app/components/tododetail/tododetail.component.ts
@@ -35,7 +35,7 @@ export class TododetailComponent implements OnInit {
       const id = this.route.snapshot.paramMap.get("listId");
       this.listSVC.getList(id).subscribe((list: any)=>{
           if(list.hasOwnProperty('obj'))
-            this.list = list.obj[0];
+            this.list = list.obj.docs[0];
       });
   }
   getTodo(){
@@ -43,51 +43,57 @@ export class TododetailComponent implements OnInit {
       const id = this.route.snapshot.paramMap.get("todoId");
       this.todoSVC.getTodo(idList, id).subscribe((todos: any)=>{
           if(todos.hasOwnProperty('obj')){
-              this.todo = todos.obj[0];
+              this.todo = todos.obj.docs[0];
               this.todo.filesOriginal = [];
-              this.getFiles(this.todo);
+              this.getFiles();
           }
       });
   }
-  updateTodo(td: ToDo): void{
-      delete td.filesOriginal;
-    this.todoSVC.editTodo(td.idList, td)
+  updateTodo(): void{
+      if(this.todo.title.trim() == "") return;
+      if(this.todo.desc.trim() == "") return;
+      this.todo.title = this.todo.title.trim();
+      this.todo.desc = this.todo.desc.trim();
+      delete this.todo.filesOriginal;
+    this.todoSVC.editTodo(this.list._id, this.todo)
         .subscribe((t: any) =>{
-            t.filesOriginal = [];
+            t.obj.filesOriginal = [];
             this.todo = t.obj;
-            this.getFiles(this.todo);
+            this.getFiles();
         });
   }
-  updateCheckBox(td: ToDo): void{
-      if(td.status)
-        td.status = false;
+  updateCheckBox(): void{
+      if(this.todo.status)
+        this.todo.status = false;
       else
-        td.status = true;
-    this.todoSVC.editTodo(this.list._id, td)
+        this.todo.status = true;
+    this.todoSVC.editTodo(this.list._id, this.todo)
         .subscribe((t: any) =>{
         });
   }
   onFileChange(file: FileList){
         this.selectedFiles = file;
   }
-  uploadFile(todo: ToDo): void{
+  uploadFile(): void{
+        if(this.todo.title.trim() == "") return;
+        if(this.todo.desc.trim() == "") return;
         this.currentFileUpload = this.selectedFiles.item(0);
-        this.FilesSVC.uploadAttachmentFile(this.currentFileUpload, todo._id)
+        this.FilesSVC.uploadAttachmentFile(this.currentFileUpload, this.todo._id)
           .subscribe((res: any)=>{
               let id = res.obj._id;
-              if(todo.files)
-                  todo.files.push(id);
+              if(this.todo.files.length > 0)
+                  this.todo.files.push(id);
               else{
-                  todo.files = [id];
+                  this.todo.files = [id];
               }
-              this.updateTodo(todo);
+              this.updateTodo();
             }, (error)=>{
                 console.log(error)
             })
   }
-  getFiles(todo: ToDo): void{
-        if(todo.files.length > 0 ){
-            this.FilesSVC.getFilesByAttachment(todo._id).subscribe((files: any)=>{
+  getFiles(): void{
+        if(this.todo.files.length > 0 ){
+            this.FilesSVC.getFilesByAttachment(this.todo._id).subscribe((files: any)=>{
                 let filesList: Array<Object> = [];
                 if(files){
                     for(let i = 0; i < files.length; i++){
@@ -103,7 +109,7 @@ export class TododetailComponent implements OnInit {
                         this.imgDisplay(obj)
                         filesList[i] = obj
                     }
-                    todo.filesOriginal = filesList;
+                    this.todo.filesOriginal = filesList;
                 }
             });
         }
@@ -118,13 +124,15 @@ export class TododetailComponent implements OnInit {
   }
   removeFile(id){
     let self = this
-    this.FilesSVC.removeFile(id).subscribe(
-        (res)=>{
-            console.log(res)
-            self.todo.files = self.todo.files.filter(f => f !== id);
-            self.updateTodo(self.todo);
-        }
-    )
+    if(this.todo.files.indexOf(id) > -1){
+        this.FilesSVC.removeFile(id).subscribe(
+            (res)=>{
+                console.log(res)
+                self.todo.files = self.todo.files.filter(f => f !== id);
+                self.updateTodo();
+            }
+        )
+    }
   }
   imgDisplay(obj){
       this.FilesSVC.downloadFile(obj.id).subscribe(
